refactor(path): reorg
