+++ fs/path.ts
@@ -1,0 +1,2 @@
+export * from "./path/mod.ts";
+export * from "./path/interface.ts";
