+++ fs/path/interface.ts
