+++ fs/path/mod.ts
