--- path/README.md
@@ -1,7 +1,0 @@
-# Deno Path Manipulation Libraries
-
-Usage:
-
-```ts
-import * as path from "https://deno.land/x/path/index.ts";
-```
