+++ test.ts
@@ -10,15 +10,15 @@ import "mkdirp/test.ts";
 import "net/bufio_test.ts";
 import "net/http_test.ts";
 import "net/textproto_test.ts";
-import "path/basename_test.ts";
-import "path/dirname_test.ts";
-import "path/extname_test.ts";
-import "path/isabsolute_test.ts";
-import "path/join_test.ts";
-import "path/parse_format_test.ts";
-import "path/relative_test.ts";
-import "path/resolve_test.ts";
-import "path/zero_length_strings_test.ts";
+import "fs/path/basename_test.ts";
+import "fs/path/dirname_test.ts";
+import "fs/path/extname_test.ts";
+import "fs/path/isabsolute_test.ts";
+import "fs/path/join_test.ts";
+import "fs/path/parse_format_test.ts";
+import "fs/path/relative_test.ts";
+import "fs/path/resolve_test.ts";
+import "fs/path/zero_length_strings_test.ts";
 import "testing/test.ts";
 
 import { runTests, completePromise } from "net/file_server_test.ts";
